Removed django-templated-mail (#18)
* Replace django-templated-mail to django send_mail
