+++ setup.py
@@ -16,7 +16,6 @@ setup(
     author_email='trench@merixstudio.com',
     install_requires=[
         'pyotp>=2.2.6',
-        'django-templated-mail>=1.1.1',
         'twilio>=6.18.1',
         'yubico-client>=1.10.0',
         'smsapi-client>=2.2.4',
