+++ trench/backends/sms_api.py
@@ -16,7 +16,7 @@ class SmsAPIBackend(AbstractMessageDispatcher):
         self.send_sms(self.to, code)
 
         return {
-            'message': _('SMS message with MFA code had been sent.')
+            'message': _('SMS message with MFA code has been sent.')
         }  # pragma: no cover # noqa
 
     def send_sms(self, user_mobile, code):
