+++ trench/settings.py
@@ -86,7 +86,7 @@ DEFAULTS = {
         'email': {
             'VERBOSE_NAME': _('email'),
             'VALIDITY_PERIOD': 60 * 10,
-            'HANDLER': 'trench.backends.templated_mail.TemplatedMailBackend',
+            'HANDLER': 'trench.backends.basic_mail.SendMailBackend',
             'SOURCE_FIELD': 'email',
         },
         'app': {
