--- trench/templates/email.html
@@ -1,3 +1,0 @@
-{% block subject %}{{ subject }}{% endblock %}
-{% block text_body %}{{ text_body }}{% endblock %}
-{% block html_body %}{{ html_body }}{% endblock %}
